backend/requirements.txt
openai
dotenv
fastapi
uvicorn
python-multipart
boto3
qrcode
pillow
supabase
pyjwt
pydantic